requirements.txt
adal<=1.2.5
applicationinsights<=0.11.9
azure-common<=1.1.25
azure-core<=1.8.2
azure-graphrbac<=0.61.1
azure-identity<=1.4.1
azure-mgmt-authorization<=0.61.0
azure-mgmt-containerregistry<=2.8.0
azure-mgmt-keyvault<=2.2.0
azure-mgmt-resource<=10.3.0
azure-mgmt-storage<=11.2.0
azure-storage-queue<=12.1.3
azureml-automl-core~=1.18.0
azureml-automl-runtime~=1.18.0
azureml-core~=1.18.0
azureml-dataprep-native<=24.0.0
azureml-dataprep-rslex<=1.2.2
azureml-dataprep<=2.4.2
azureml-dataset-runtime[fuse,pandas]~=1.18.0
azureml-defaults~=1.18.0
azureml-interpret~=1.18.0
azureml-model-management-sdk<=1.0.1b6.post1
azureml-pipeline-core~=1.18.0
azureml-pipeline-steps~=1.18.0
azureml-telemetry~=1.18.0
azureml-train-automl-client~=1.18.0
azureml-train-automl-runtime~=1.18.0
azureml-train-automl~=1.18.0
azureml-train-core~=1.18.0
azureml-train-restclients-hyperdrive~=1.18.0
backports.tempfile<=1.0
backports.weakref<=1.0.post1
boto3<=1.15.18
boto<=2.49.0
botocore<=1.18.18
certifi<=2020.6.20
cffi<=1.14.3
chardet<=3.0.4
click<=7.1.2
cloudpickle<=1.6.0
configparser<=3.7.4
contextlib2<=0.6.0.post1
cryptography<=3.2.1
cython<=0.29.14
dill<=0.3.3
distro<=1.5.0
docker<=4.3.1
dotnetcore2<=2.1.17
fire<=0.3.1
flask<=1.0.3
fusepy<=3.0.1
gensim<=3.8.3
gunicorn<=19.9.0
idna<=2.10
importlib-metadata<=2.0.0
interpret-community<=0.15.3
interpret-core[required]<=0.2.1
isodate<=0.6.0
itsdangerous<=1.1.0
jeepney<=0.4.3
jinja2<=2.11.2
jmespath<=0.10.0
joblib<=0.14.1
json-logging-py<=0.2
jsonpickle<=1.4.1
keras2onnx<=1.6.0
liac-arff<=2.5.0
lightgbm<=2.3.0
markupsafe<=1.1.1
msal-extensions<=0.2.2
msal<=1.6.0
msrest<=0.6.19
msrestazure<=0.6.4
ndg-httpsclient<=0.5.1
nimbusml<=1.8.0
numpy<=1.18.5
oauthlib<=3.1.0
onnx<=1.7.0
onnxconverter-common<=1.6.0
onnxmltools<=1.4.1
onnxruntime<=1.3.0
packaging<=20.4
pandas<=0.25.3
pathspec<=0.8.0
patsy<=0.5.1
pmdarima<=1.1.1
portalocker<=1.7.1
protobuf<=3.13.0
psutil<=5.7.3
py-cpuinfo<=5.0.0
pyarrow<=1.0.1
pyasn1<=0.4.8
pycparser<=2.20
pyjwt[crypto]<=1.7.1
pyopenssl<=19.1.0
pyparsing<=2.4.7
python-dateutil<=2.8.1
pytz<=2020.4
pywin32<=227
requests-oauthlib<=1.3.0
requests<=2.24.0
ruamel.yaml.clib<=0.2.2
ruamel.yaml<=0.16.12
s3transfer<=0.3.3
scikit-learn<=0.22.2.post1
scipy<=1.5.2
secretstorage<=3.1.2
shap<=0.34.0
six<=1.15.0
skl2onnx<=1.4.9
sklearn-pandas<=1.7.0
smart-open<=1.9.0
statsmodels<=0.10.2
termcolor<=1.1.0
tqdm<=4.51.0
typing-extensions<=3.7.4.3
urllib3<=1.25.11
websocket-client<=0.57.0
werkzeug<=1.0.1
zipp<=3.4.0